wl12xx: change packet retry count
default retry number was 10. we found out that in noisy environment
there were case we were failed to send the packet.
in that case it took time until the upper layer retransmit the
packet which cause degradation in TCP TP.
change the default to 50 in order to solve this problem

Signed-off-by: Yossi Wortzel <yossiw@ti.com>